+++ tests/Enums/IntegerValues.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace BenSampo\Enum\Tests\Enums;
+
+use BenSampo\Enum\Enum;
+
+final class IntegerValues extends Enum
+{
+    const B = 1;
+    const A = 2;
+}